Update README.md
Data science has become a critical component of modern-day businesses, to handle and analyze data. Python is a powerful programming language used widely in data science and offers several libraries and tools to perform complex data analysis tasks.
In this workshop, we will introduce you to the fundamental concepts of Python programming for data science. You will learn the essential techniques for data exploration, cleaning, analysis, and visualization using Python libraries like Pandas, NumPy, Matplotlib, and Seaborn. We will also cover supervised machine learning algorithms using popular Python libraries like Scikit-learn. The workshop is structured into eight one-hour sessions that provide an understanding of the key concepts and hands-on experience of using Python for data science.
By the end of the workshop, you will be equipped with the skills and knowledge to handle and analyze data using Python. You will have the opportunity to apply concepts hands-on learned throughout the workshop. Each session is expected to be 1 hour long. Participants are not required to have an advanced knowledge of programming or mathematics, but a basic understanding is desirable. This workshop emphasizes more on application of Python in data science rather than programming. Understanding of software tools for coding will be provided to enable participants implement the hands-on part and gain proficiency in this exciting field. The course outline is subject to change based on the pace of the workshop and the requirements of the participants.